Move the test suite into rust-app, hook into cargo.
This requires a bit of hackery with wrapping speculos, but much less
than I was concerned might be the case. We can't do this by just having
tests/ts-tests.rs run the suite directly because we don't want to do a
multitarget build, so tests/ts-tests.rs actually gets built _for the
ledger_, which doesn't get much access to system calls.