Add a prompt for cross-chain transfer, and disable the warning when used
This doesn't include any provision for streaming longer code matches in
the future; it simply buffers 600 characters of code and then does some
string slice manipulation to check if it is exactly the cross-chain
transfer produced by the web signer tool.

Also includes a small optimization on the way for memory use of
capability parameters, which were costing roughly double memory; this
unfortunately is not enough to allow full 256-byte address names.