Fix bug in public key response
THe total pubkey is:

    1 byte (y mod 2) || 32 bytes x

We skipped the parity but *didn't* correspondingly decrement the length
prefix we sent over. This was previously compensated by the off-by-one
reading.

Now the reading side is fixed, and this is fixed too, by computing the x
coordinate slice, and then simply and stupidly sending it's length and
it and it over.